+++ release/utils/index.js
@@ -7,4 +7,5 @@ require('./setOverflow/index.js');
 require('./closeRouterModal/index.js');
 require('./withModal/index.js');
 require('./actionProp/index.js');
+require('./createFormDate/index.js');