+++ src/ui/forms/Form/types.ts
@@ -5,16 +5,19 @@ import { State } from 'watch-state'
 import { FormContext } from '../../../hooks'
 import { FlexProps } from '../../layout'
 
+export type FormMethod = 'GET' | 'HEAD' | 'POST' | 'DELETE' | 'PUT' | 'CONNECT' | 'OPTIONS' | 'TRACE' | 'PATCH'
+
 export interface FormProps extends FlexProps<HTMLFormElement> {
   action?: WatchProp<string>
   loading?: State<boolean>
   notification?: string
+  method?: FormMethod
   onerror?: (form: FormContext, error?: any) => void
   onsuccess?: (form: FormContext) => void
 }
 
 export type FormErrorHandle = (error: ValidationError<any>, form: FormContext) => string | Promise<string>
-export type FormActionHandle = (action: string, form: FormContext) => Promise<any> | void
+export type FormActionHandle = (action: string, form: FormContext, method: FormMethod) => Promise<any> | void
 
 export interface FormFieldProps <V = string> {
   name: string